Check if password fields match
Using JS I prevented the submission of the form if the confirm
password doesn't match the password. I also removed the valid
and invalid psuedo classes. It's not a good validation method
from a UX point of view. Required elements are by default
invalid when empty, and the result was that the label turned
red when typing in.